XXX more regex: Add ANYOFHr regnode
This commit adds a lower bound for the first UTF-8 byte matchable by an
ANYOFH node.  The flags field is otherwise unused, and using it for this
purpose allows code to rule out match possibilities without having to
convert from UTF-8 to code point.

It might be better to do the inverse instead, to have the field be an
upper bound.  The reason is that the conversion is cheap for smaller
numbers.  The commit following mostly addresses this.

This commit adds a new regnode, ANYOFHr, like ANYOFH, but it also has a
loose upper bound for the first UTF-8 byte matchable by the node.  It
would be nice to have a tight upper bound, but to do so requires 4 more
bits than are available without changing the node arguments types.
Having a loose bound is better than no bound, and comes essentially
free, by using two bits in the current ANYOFH node that are redundant,
and requiring only a few extra, pipelined, mask, etc instructions at run
time, no extra conditionals.  Any ANYOFH nodes that would benefit from
having an upper bound will instead become this node type.

Its used is based on the following observations.

There are 64 possible start bytes, so the full range can be expressed in
6 bits.  This means that the flags field in ANYOFH nodes containing the
start byte has two redundant bits that can be used for something else.

An ANYOFH node only happens when there is no matching code point in the
bit map.so the smallest code point that could be is 256.  The start byte
for that is C4, so there are actually only 60 possible start bytes.
(perl can be compiled with a larger bit map in which case the minimum
 start byte would be even higher.)

A second observation is that knowing the highest start byte is above F0
is no better than knowing it's F0.  This is because the highest code
point whose start byte is F0 is U+3FFFF, and all code points above that
that are currently allocated are all very specialized.  And there's no
likelihood of that changing anytime soon as there's plenty of
unallocated space below that.  So if an ANYOFH node's highest start byte
is F0 or above, leave it ANYOFH, and no harm down.

That means the highest start byte we care about in ANYOFHr is EF.  That
cuts the number of start bytes we care about down to 43, still 6 bits
required to represent them, but it allows for the following scheme

Populate the flags field by subtracting C4 from the lowest start byte
and shift left 2 bits.  That leaves the the bottom two bits unused.
We use them as follows, where x is the start byte of the lowest code
point in the node:

11  The upper limit of the range can be as much as (EF - x) / 8
10  The upper limit of the range can be as much as (EF - x) / 4
01  The upper limit of the range can be as much as (EF - x) / 2
00  The upper limit of the range can be as much as  EF

The possible high bytes for C4 are C9 CE D9 EF
The possible high bytes for D0 are D3 D7 DF EF
The possible high bytes for E0 are E1 E3 E7 EF

Thus we have provided a loose upper bound using two bits.  An alternate
scheme could have had the intervals all the same, but this provides a
tighter bound if the range is small, which...

For EBCDIC the range is is from C8 to F4,
l