Add 'C' flag to embed.fnc, for "core-only" fcns
Previously the x flag was (mis-)used for this purpose.  Some functions
that are accessible by XS writers shouldn't be used by them, such as
helper functions that an inline public function calls.  This new flag
alows that to be specified, without having to mark the function as
experimental, which could be misleading to later core maintainers